+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- VIMEET-2154 - securize impersonation
+
+### Added
+- VIMEET-2157 - Save import mapping
 
 ## [2.37.0] - 2020-07-06
 ### Added
@@ -17,9 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MV-243 - Add media link in sheet pdf printing
 ### Added
 - VIMEET-2149 - Change Admin password algorithm to argon2i and increase length
-
-### Added
-- VIMEET-2157 - Save import mapping
 
 ## [2.35.1] - 2020-06-30
 ### Fixed